+++ src/FluxFlow.Components.Control.Composition/ControlCompositionNodeTypes.cs
@@ -1,8 +1,0 @@
-namespace FluxFlow.Components.Control.Composition;
-
-public static class ControlCompositionNodeTypes
-{
-    public const string Filter = "flow.filter";
-
-    public const string When = "flow.when";
-}