+++ src/FluxFlow.Components.Routing.Composition/RoutingCompositionResourceNames.cs
@@ -4,8 +4,6 @@ public static class RoutingCompositionResourceNames
 {
     public const string Clock = "clock";
 
-    public const string RouteKeySelector = "routeKeySelector";
-
     public const string KeySelector = "keySelector";
 
     public const string SideSelector = "sideSelector";