+++ src/FluxFlow.Components.Routing/RoutingComponentPorts.cs
@@ -8,10 +8,8 @@ public static class RoutingComponentPorts
     public const string Left = "Left";
     public const string Right = "Right";
     public const string Output = "Output";
-    public const string Routed = "Routed";
     public const string Result = "Result";
     public const string Matched = "Matched";
     public const string Timeouts = "Timeouts";
-    public const string Default = "Default";
     public const string Errors = "Errors";
 }